backend/requirements-vendor.txt
Django==1.8
dj-database-url==0.3.0
django-filter==0.8
djangorestframework==3.1.1
bleach==1.4
psycopg2
